+++ __tests__/template-string.ts
@@ -1,0 +1,10 @@
+import { splitByCommasWithEscapes } from "../src/utils";
+
+
+describe("template-string", () => {
+  it("can split commas with escapes", () => {
+    const values = splitByCommasWithEscapes("weapon, off-hand, acc1");
+
+    expect(values).toEqual(["weapon", "off-hand", "acc1"]);
+  });
+});
